% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_estimation_functions.R
\name{aggregate_strata}
\alias{aggregate_strata}
\title{Aggregate strata using specified weighting and calculate associated
weighted-average Standard Error estimates as well.}
\usage{
aggregate_strata(gsum, weighting = c("normal", "double", "precision"))
}
\arguments{
\item{gsum}{Data table of strata-level estimates (point estimates
and standard errors, etc)}

\item{Weighting:}{normal means weight by estimated number of
compliers, double means square of that, precision means 1/SE^2.}
}
\value{
Aggregated result
}
\description{
Aggregate strata using specified weighting and calculate associated
weighted-average Standard Error estimates as well.
}
