% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_simulation_functions.R
\name{default_sim_params}
\alias{default_sim_params}
\title{Default simulation parameters for simulation code}
\usage{
default_sim_params()
}
\value{
Small tibble with simulation parameters that can be passed
to simulation functions.
}
\description{
Default simulation parameters for simulation code
}
