% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_data_generation_functions.R
\name{describe_sim_data}
\alias{describe_sim_data}
\title{Describe a dataset with compliance}
\usage{
describe_sim_data(df)
}
\value{
Dataframe with several summary measures.  sd0, the variance of the Y0
in the dataframe.
}
\description{
Given a dataframe that has both potential outcomes (e.g., a simulated
complete dataset), calculate some summary measures of interest.
}
