% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_estimation_functions.R
\name{IV.est.oracle}
\alias{IV.est.oracle}
\title{Parameter calculation and oracle estimation (For simulation studies)}
\usage{
IV.est.oracle(data, include_est = FALSE)
}
\arguments{
\item{data}{Dataframe with named Y0, Y1, S0, S1 and Z columns.}
}
\description{
This function does two things:
}
\details{
(1) Calculate some simulation statistics (sample parameter values) This is
done with the full schedule of potential outcomes, something we would not see
empirically.

(2) (If include_est = TRUE) estimate (using an oracle estimator) the CACE.
The oracle estimator is the difference in means on the compilers only.

(this method is for simulation code).
}
