% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_estimation_functions.R
\name{IV.est}
\alias{IV.est}
\title{Implement different versions of classic IV estimation.}
\usage{
IV.est(data, Yobs = "Yobs", S = "S", Z = "Z", tolerance = 1/10000)
}
\arguments{
\item{Yobs, }{S, Z : column names of outcome, treatment receipt and
treatment assignment (instrument), respectively.}
}
\value{
One row tibble of results.  If not both treatment and
control units present, return tibble with NAs for estimated
impacts, etc.
}
\description{
Implement different versions of classic IV estimation.
}
