% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_estimation_functions.R
\name{IV.est.strat.cont}
\alias{IV.est.strat.cont}
\title{Estimate stratifying on a continuous variable X}
\usage{
IV.est.strat.cont(df, strat_var = "X")
}
\arguments{
\item{df}{Dataframe with standard outcome columns.}

\item{strat_var}{Name of column in data that holds a continuous covariate.}
}
\description{
This just cuts provided continuous covariate X into 4 bins and does the post-stratification on the
resulting categorical variable.
}
