% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_estimation_functions.R
\name{IV.est.strat}
\alias{IV.est.strat}
\title{This will group units based on X and then estimate the LATE within
each subgroup. It then averages the subgroups with weight
proportional to the estimated proportion of compliers in each
group.}
\usage{
IV.est.strat(
  data,
  Yobs = "Yobs",
  S = "S",
  Z = "Z",
  strat_var = "X",
  DSS_cutoff = 0.02,
  drop_without_warning = FALSE,
  include_blocks = TRUE,
  include_FSS = TRUE,
  return_strata_only = FALSE
)
}
\arguments{
\item{Yobs, }{S, Z : column names of outcome, treatment receipt and
treatment assignment (instrument), respectively.}

\item{strat_var}{Name of column in data that holds a categorical
variable that we should stratify on.}

\item{drop_without_warning}{If TRUE will drop blocks that are
inestimable due to not having at least 2 tx and 2 co units.  If
FALSE will drop, but also throw warning.}
}
\description{
STRAT_cw: Estimate LATE in each strata, then take the weighted
mean, typically weighting by estimated number of compliers.
}
\details{
STRAT2: Estimate the post-stratified ITT and pi, take the ratio to
get overall LATE estimate.
}
