% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_data_generation_functions.R
\name{make.dat.tuned}
\alias{make.dat.tuned}
\title{Make stratified noncompliance dataset}
\usage{
make.dat.tuned(
  N,
  pstrat = rep(0.25, 4),
  pi_c = c(0, 0.01, 0.12, 0.3),
  pi_n = 1 - pi_c,
  Ybar0 = c(2, 5, 7, 10),
  sd0 = 3,
  tau = c(0, 1, 2, 3),
  at_shift = -1,
  nt_shift = 1,
  exclusion_impact = 0,
  maintain_tau = TRUE
)
}
\arguments{
\item{pstrat}{Proportion of units in each strata.}

\item{pi_c}{Proportion of compliers in each strata.}

\item{pi_n}{Proportion of never-takers in each stata, defaults to
remainder making no always-takers.}

\item{Ybar0}{baseline mean outcome of each strata.}

\item{sd0}{Within-group individual residual variance.}

\item{tau}{Complier average causal effect for each strata.}

\item{at_shift}{Always-takers will have their potential outcomes
shifted by this much, relative to compliers.}

\item{nt_shift}{Never-takers will have their potential outcomes
shifted by this much, relative to compliers.}
}
\value{
Dataframe of individual level data.  X denotes strata
membership (categorical).
}
\description{
Make dataset with different strata with different numbers of
compliers in each strata.
}
