% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_data_generation_functions.R
\name{make_dat}
\alias{make_dat}
\title{Simulate data for non-compliance in an RCT}
\usage{
make_dat(
  N,
  pi_c = NULL,
  frac_nt = 1/3,
  one_sided = TRUE,
  at_shift = 0,
  nt_shift = 0,
  sd0 = 1,
  pred_comp = "yes",
  pred_Y = "yes",
  het_tx = "yes",
  exclusion_impact = 0,
  perfect_X = FALSE,
  scaled_C = NULL,
  verbose = FALSE,
  pstrat = rep(0.25, 4),
  pi = c(0.01, 0.05, 0.55, 0.95),
  Ybar0 = c(4, 3, 2, 1),
  tau = c(0, 1, 2, 3),
  params = NULL,
  include_POs = TRUE
)
}
\arguments{
\item{pi_c}{Overall proportion of compliers}

\item{frac_nt}{The fraction of non-compliers that are never takers
(can vary by strata).}

\item{one_sided}{If FALSE, split non-compliers into AT and NT.}

\item{nt_shift}{shifts the never takers mean outcome so they are
different from compliers.}

\item{sd0}{baseline Y0 standard deviation not accounting for shift
of NT and AT and treatment impacts.}

\item{pred_comp}{= FALSE means all strata have same proportion of
compliers}

\item{pred_Y}{= FALSE means all strata have same Y0 means (for
compliers).}

\item{exclusion_impact}{Treatment impact for noncompliers (i.e.
violation of the exclusion restriction).}

\item{perfect_X}{= TRUE means make X4 all compliers, and X1, X2,
and X3 random mix of the noncompliers.}

\item{scaled_C}{If a number, pack compliers into the strata with
proportions equal to a geometric series with, for X4, X3, X2 and
X1 relative proportion 1, scaled_C, scaled_C^2, scaled_C^3 such
that the sum equals the target desired pi_c.}

\item{pstrat}{The relative sizes of strata from X1 to X4.
Overridden if params is not null}

\item{pi}{The relative proportion of compliers in each strata (this
will be rescaled to achieve overall pi_c proportion compliers).}

\item{params}{Dataframe with pstrat, pi, Ybar0, and tau as default
values for the strata.  Will overwrite the passed values of
these, if params is not null.}

\item{include_POs}{Should the potential outcomes be included in the
final returned data?  If FALSE, then you get data that would look
like what you would get in the field.  TRUE means you have the
underlying truth to calculate actual (oracle) effects.}
}
\description{
Make data given set parameters.  Idea is to have standardized
outcomes regardless of parameters, mostly.
}
