% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_simulation_functions.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation of noncompliance in an RCT}
\usage{
run_sim(
  reps,
  N,
  nt_shift,
  sd0,
  pi_c,
  pred_comp,
  pred_Y,
  het_tx,
  one_sided = TRUE,
  scaled_C = NULL,
  perfect_X = FALSE,
  shuffle = FALSE,
  params = default_sim_params(),
  seed = NULL,
  ...
)
}
\arguments{
\item{reps}{Number of replications}

\item{N}{Number of observations}

\item{nt_shift}{Shift in the mean of the outcome for the
never-taker group.}

\item{sd0}{Standard deviation of the outcome}

\item{pi_c}{Proportion of compliers}

\item{pred_comp}{Is the covariate predictive of compliance?}

\item{pred_Y}{Is the covariate predictive of the outcome?}

\item{het_tx}{Is the treatment effect heterogeneous?}

\item{one_sided}{Is compliance one-sided (TRUE) or two-sided
(FALSE)?}

\item{scaled_C}{passed to DGP function}

\item{perfect_X}{passed to DGP function}

\item{shuffle}{Should the X variable be shuffled to break all
connection to anything else?}

\item{params}{Simulation parameters as a tibble (see, e.g.,
default_sim_params()).}

\item{seed}{Seed for the random number generator}

\item{...}{Additional arguments passed to the DGP function}
}
\value{
A tibble with the results of the simulation.
}
\description{
Run a simulation of noncompliance in an RCT
}
