% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_data_generation_functions.R
\name{summarize_sim_data}
\alias{summarize_sim_data}
\title{Summarize simulated dataset}
\usage{
summarize_sim_data(dd)
}
\arguments{
\item{dd}{Dataframe of simulation data with three columns: Yobs,
complier (a flag of being a complier) and X (a covariate).}
}
\value{
Dataframe with R2_y and R2_comp from two regressions.
}
\description{
Calculate how predictive the covariate X is in terms of outcome and of
compliance.
}
